% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Génère un rapport pour une commune et un département spécifiés.}
\usage{
generer_rapport(commune, departement, output)
}
\arguments{
\item{commune}{Une chaîne de caractères ou un vecteur de communes pour lesquelles générer le rapport.}

\item{departement}{Une chaîne de caractères ou un vecteur de départements pour lesquels générer le rapport.}

\item{output}{Le chemin complet du fichier où le rapport sera enregistré.}
}
\value{
Aucun retour. Le rapport est généré et sauvegardé à l'emplacement spécifié.
}
\description{
Génère un rapport pour une commune et un département spécifiés.
}
